+++ build-config.ts
@@ -20,18 +20,30 @@ export const webComponentList: WebComponentBuildConfig[] = [
     outputPath: "./i18n/dist/index.js",
     umdName: "JBCoreI18N",
     dir: "./i18n",
-  }
+  },
 ];
 export const reactComponentList: ReactComponentBuildConfig[] = [
+  {
+    name: "jb-core-i18n-react",
+    path: "./i18n/react/lib/index.ts",
+    outputPath: "./i18n/react/dist/index.js",
+    umdName: "JBCoreI18NReact",
+    external: ["react", "jb-core/i18n"],
+    dir: "./i18n/react",
+    globals: {
+      react: "React",
+      "jb-core/i18n": "JBCoreI18N",
+    },
+  },
   {
     name: "jb-core-react",
     path: "./react/lib/index.ts",
     outputPath: "./react/dist/index.js",
     umdName: "JBCoreReact",
-    external:['react'],
-    dir:"./react",
-    globals:{
-      'react':"React"
-    }
-  }
+    external: ["react"],
+    dir: "./react",
+    globals: {
+      react: "React",
+    },
+  },
 ];