+++ i18n/lib/index.ts
@@ -1,2 +1,2 @@
-export * from "./i18n.js"
-export * from './dictionary.js'
+export * from "./i18n.js";
+export * from "./dictionary.js";