+++ i18n/react/lib/index.ts
@@ -1,0 +1,1 @@
+export * from "./use-jb-i18n.js";